ansible/roles/flask-app/files/requirements.txt
flask==3.0.0
dnspython==2.4.2
requests==2.31.0
cryptography==41.0.7
pyopenssl==23.3.0
schedule==1.2.0
redis==5.0.0
rq==1.15.1
psycopg2-binary==2.9.9
python-whois==0.8.0
bcrypt==4.1.2
pyjwt==2.8.0
boto3==1.34.0
flask-cors==4.0.0
python-dotenv==1.0.0
geoip2==4.7.0
stripe==7.4.0

# WebSocket Support
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3

# GraphQL (compatible versions)
graphene==2.1.9
flask-graphql==2.0.1
graphql-core==2.3.2

# Additional dependencies for WebSocket/GraphQL
hiredis==2.2.3
gevent==23.9.1
